Remove ref to `bytes` return type
Future versions will not provide a built-in bytes converter anymore (though it would be easy for a caller to implement).
The rationale is that (we think) the pixel data is rarely needed as bytes object itself, so providing such a converter would only encourage bad use of an intermediate copy. Converting from a c_ubyte array directly to the target object is more efficient and should be possible instead.